feat(verl): graduate rollout-gateway backend
