+++ docs/site/src/content/docs/blog/training-multi-step-agents-rl.mdx
@@ -380,7 +380,7 @@ Here are a few of our experimental findings:
   and the accompanying write-up,
   [Training a code-migration agent on AgentCore](https://rllm-project.com/post.html?post=agentcore_migrationbench.md).
 - **OfficeBench** — see the training code at
-  [agentcore-rl-toolkit `verl/examples/officebench_agent`](https://github.com/awslabs/agentcore-rl-toolkit/tree/main/src/agentcore_rl_toolkit/backends/verl/examples/officebench_agent).
+  [agentcore-rl-toolkit `strands_officebench_agent`](https://github.com/awslabs/agentcore-rl-toolkit/tree/main/examples/strands_officebench_agent).
 
 ## What's next
