--- docs/site/src/content/docs/troubleshooting/verl.md
@@ -1,14 +1,0 @@
----
-title: verl troubleshooting
-description: Known issues and fixes when training with the verl (FSDP) backend.
----
-
-:::caution[Preview]
-The verl integration is under active development. This page
-collects known issues as they surface. Setup guide:
-[verl](/agentcore-rl-toolkit/guides/verl-backend-setup/).
-:::
-
-*No issues reported yet.* Please file a GitHub issue on
-[awslabs/agentcore-rl-toolkit](https://github.com/awslabs/agentcore-rl-toolkit/issues)
-with repro steps if you hit one.
