--- src/agentcore_rl_toolkit/backends/experimental/verl/examples/migration_agent/README.md
@@ -1,40 +1,0 @@
-# MigrationBench agent (Megatron + LoRA)
-
-GRPO for the Java 8 to 17
-[`strands_migration_agent`](../../../../../../../examples/strands_migration_agent),
-using Qwen3-Coder-30B-A3B with Megatron expert parallelism and LoRA.
-
-## Install
-
-Megatron-Bridge 0.5.0 requires Python 3.12:
-
-```bash
-uv venv --python 3.12
-uv sync --extra verl-experimental --group verl-experimental-megatron
-```
-
-## Prepare data
-
-Run these commands from this directory. The first command prepares repository
-tarballs and metadata in S3; the second creates the payload-only parquet files used
-by verl.
-
-```bash
-python ../../../../../../../examples/strands_migration_agent/preprocess.py \
-  --s3-bucket-name <data-bucket>
-python preprocess_migrationbench.py --s3-bucket-name <data-bucket>
-```
-
-## Train
-
-```bash
-export AGENT_RUNTIME_ARN=arn:aws:bedrock-agentcore:...:runtime/...
-export ACR_S3_BUCKET=your-results-bucket
-wandb login
-./megatron_lora_sync_grpo.sh
-```
-
-The per-turn generation limit is configured in `agentcore_agent.yaml`; the shell
-script separately configures the cumulative model context and verl storage limits.
-Qwen3-Coder's chat-template hash maps to the gateway's `qwen3_5` response schema for
-tool-call parsing.
