+++ tests/backends/verl/test_gateway_host.py
@@ -5,7 +5,7 @@ import urllib.request
 
 import pytest
 
-from agentcore_rl_toolkit.backends.experimental.verl.gateway_host import _url_host, get_or_start_gateway
+from agentcore_rl_toolkit.backends.verl.gateway_host import _url_host, get_or_start_gateway
 
 from .conftest import FakeLLMServerClient, FakeTokenizer
